Adjustments
Query, Services configurations, Response case, route & etc.

+++ Samples/StoreAPI/StoreAPI.Core.Application/Products/Queries/GetProductsByFilter/GetProductsByFilterQueryHandler.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using ModelWrapper.Extensions;
 using StoreAPI.Core.Application.Interfaces.Infrastructures.Data;
 using StoreAPI.Core.Domain.Entities;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -18,8 +19,9 @@ namespace StoreAPI.Core.Application.Products.Queries.GetProductsByFilter
 
         public async Task<GetProductsByFilterQueryResponse> Handle(GetProductsByFilterQuery request, CancellationToken cancellationToken)
         {
-            int resultCount = 0;
-            var data = await Context.Products.FullSearch(request, ref resultCount)
+            long resultCount = 0;
+            var data = await Context.Products
+                .FullSearch(request, ref resultCount)
                 .AsNoTracking()
                 .ToListAsync(cancellationToken);
